Refactor _backend module
